Made Queuesim a user defined package. Further reorganization of unit test.
