+++ Queuesim/TrafficGenerator.py
@@ -1,10 +1,11 @@
-from Customer import Customer
+from ServerBase import ServerBase
 
 class TrafficGenerator:
     pass
 
 import unittest
 class testTrafficGenerator:
+    pass
 
-if __name__ == 'main'
+if __name__ == 'main':
     unittest.main()
