+++ Queuesim/testTrafficUnit.py
@@ -1,31 +1,18 @@
-from ServerBase import ServerBase
-
-class TrafficUnit:
-    def __init__(self, server):
-        self.__action__ = server.action()
-
-    def response_time(self):
-        return self.__action__.value
-
 import unittest
 import simpy
 import numpy as np
 
+from TrafficUnit import TrafficUnit
+from MockTestServer import MockTestServer
+
 class testTrafficUnit(unittest.TestCase):
     class __Simulation__:
-        class __TestServer__(ServerBase):
-            def __init__(self, env, resource):
-                super().__init__(env, resource)
-
-            def service_time(self):
-                return 5
-
         def __init__(self):
             self.__traffic__ = []
             env = simpy.Environment()
             resource = simpy.Resource(env, capacity=1)
             env.process(self.__traffic_source__())
-            self.__server__ = self.__TestServer__(env, resource)
+            self.__server__ = MockTestServer(env, resource)
             self.__env__ = env
 
         def __traffic_source__(self):
@@ -48,6 +35,3 @@ class testTrafficUnit(unittest.TestCase):
         simulation = self.__Simulation__()
         simulation.run()
         np.testing.assert_almost_equal(simulation.response_times(), [5, 7, 10])
-
-if __name__ == '__main__':
-    unittest.main()
